requirements.txt
Django==4.2.3
wheel
psycopg2
django_compressor
django-libsass
django-summernote
Pillow
django-imagekit